pyproject.toml
[tool.poetry]
name = "chrome-ext-api"
version = "0.1.0"
description = "An api for storing and fetching video files recorded by a chrome extension. HNGX stage 5 task"
authors = ["Nasir Abdulrasheed <mypythtesting@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
pymongo = "^4.5.0"
python-decouple = "^3.8"
python-multipart = "^0.0.6"
uvicorn = {extras = ["main"], version = "^0.23.2"}
fabric2 = "^3.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
